+++ src/data/writeToStore.ts
@@ -59,6 +59,16 @@ class WriteError extends Error {
   public type = 'WriteError';
 }
 
+function enhanceErrorWithDocument(error: Error, document: DocumentNode) {
+  // XXX A bit hacky maybe ...
+  const enhancedError = new WriteError(`Error writing result to store for query ${
+    document.loc && document.loc.source && document.loc.source.body
+  }`);
+  enhancedError.message += '/n' + error.message;
+  enhancedError.stack = error.stack;
+  return enhancedError;
+}
+
 /**
  * Writes the result of a query to the store.
  *
@@ -100,18 +110,22 @@ export function writeQueryToStore({
 
   variables = assign({}, getDefaultValues(queryDefinition), variables);
 
-  return writeSelectionSetToStore({
-    dataId: 'ROOT_QUERY',
-    result,
-    selectionSet: queryDefinition.selectionSet,
-    context: {
-      store,
-      variables,
-      dataIdFromObject,
-      fragmentMap,
-      fragmentMatcherFunction,
-    },
-  });
+  try {
+    return writeSelectionSetToStore({
+      dataId: 'ROOT_QUERY',
+      result,
+      selectionSet: queryDefinition.selectionSet,
+      context: {
+        store,
+        variables,
+        dataIdFromObject,
+        fragmentMap,
+        fragmentMatcherFunction,
+      },
+    });
+  } catch (e) {
+    throw enhanceErrorWithDocument(e, query);
+  }
 }
 
 export type WriteContext = {
@@ -161,11 +175,7 @@ export function writeResultToStore({
       },
     });
   } catch (e) {
-    // XXX A bit hacky maybe ...
-    const e2 = new Error(`Error writing result to store for query ${document.loc && document.loc.source && document.loc.source.body}`);
-    e2.message += '/n' + e.message;
-    e2.stack = e.stack;
-    throw e2;
+    throw enhanceErrorWithDocument(e, document);
   }
 }